requirements.txt
Flask==2.0.1
numpy==1.19.5
Pillow==8.3.2
requests==2.25.1
urllib3==1.26.3
azure-common==1.1.27
azure-core==1.18.0
azure-storage-blob==12.9.0
azure-storage-common==2.1.0
Werkzeug==2.0.1
